fix: microbatch and scheduler fixes
