+++ src/scheduler/api/scheduler_run.py
@@ -3,6 +3,11 @@ import datetime
 from django.db import transaction
 from django.http import HttpResponse
 from django.urls import reverse
+from job_manager.models.task import Task
+from job_manager.utils import (
+    consolidate_job_datetimes,
+    consolidate_parent_task_datetimes,
+)
 
 from ..models import (
     ResourceAllocations,
@@ -74,7 +79,9 @@ def start_scheduler_run(request):
     """
     scheduler_start_time = datetime.datetime.now(datetime.timezone.utc)
     try:
-        scheduled_task = SchedulingService(horizon_weeks=5).run()
+        scheduled_task = SchedulingService(horizon_weeks=5, user=request.user).run(
+            selected_tasks=Task.objects.filter(sub_tasks__isnull=True),
+        )  # Run the scheduler for the microbatched subtasks
         scheduler_end_time = datetime.datetime.now(datetime.timezone.utc)
 
         # check if the scheduleed task has error key
@@ -105,6 +112,9 @@ def start_scheduler_run(request):
             scheduler_end_time,
             scheduler_status,
         )
+        
+        consolidate_parent_task_datetimes()  # Updates the parent Task datetimes based on the Subtask start/end datetimes
+        consolidate_job_datetimes()  # Updates the Job datetimes based on the new parent Task datetimes
 
         if request.htmx:
             headers = {"HX-Redirect": reverse("scheduler_runs")}
