+++ src/templates/objects/list.html
@@ -362,10 +362,6 @@
         <button class="text-base font-semibold text-white px-6 py-3 bg-[#023E8A] rounded-md"
             hx-post="{% url 'generate_task_flows' %}" type="button">Generate Task Flows</button>
         {% endif %}
-        {% if model_name == 'microbatch_flows' %}
-        <button class="text-base font-semibold text-white px-6 py-3 bg-[#023E8A] rounded-md"
-            hx-post="{% url 'run_microbatch_scheduling' %}" type="button">Run Microbatch Scheduling</button>
-        {% endif %}
 
         <a href="{% url model_name|add:'_form' %}"
             class="text-base font-semibold text-white px-6 py-3 bg-[#023E8A] rounded-md">Create New {{ model_title | title }}</a>
